+++ extensions/vercast/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Vercast Changelog
 
+## [2.16] - 2026-02-15
+
+- Added `Search AI Gateway Logs` command
+
+## [2.15] - 2026-01-29
+
+- Added `Search AI Gateway Models` command
+
 ## [2.14] - 2026-01-20
 
 - Removed token length validation as vercel tokens can now be longer